pyproject.toml
[tool.poetry]
name = "xrpld-netgen"
version = "1.10.0"
description = "Xrpld network generator"
authors = ["Denis Angell <dangell@transia.co>"]

[tool.poetry.dependencies]
python = "^3.9.6"
pyyaml = "^6.0.1"
requests = "^2.31.0"
xrpld-publisher = "^2.0.0"
xrpl-helpers = "^1.0.4"

[tool.poetry.dev-dependencies]
pytest = "^7.2.1"
flake8 = "^3.8.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
xrpld-netgen = "xrpld_netgen.cli:main"
